requirements.txt
chromadb
langchain
sentence-transformers
unstructured
python-dotenv
langchain-chroma>=0.1.2
langchain-community
langchain-huggingface
fastapi
uvicorn
langchain-groq
pi-heif
"unstructured[inference,pdf]" 
fitz 
pillow 
opencv-python 
PyPDF2
streamlit